fix(agent,windows): ship breeze-user-helper with asInvoker manifest (#949 regression)
breeze-user-helper.exe was built from ./cmd/breeze-agent, so Go auto-linked the
agent's requireAdministrator manifest (added in #949 for the #944 MSI VERSIONINFO
fix) into it. The sessionbroker spawns the user-helper into the logged-in user's
session with that user's token; under UAC a normal admin's token is filtered
(non-elevated), so CreateProcessAsUser fails with ERROR_ELEVATION_REQUIRED (740),
breaking run-as-user scripts and the user-context Helper launch on every endpoint
whose console user isn't the built-in Administrator. The role=system spawn is
unaffected (SYSTEM token is elevated), which masked the bug; our Windows test VM
logs in as the built-in Administrator (full token), so it never reproduced.

Fix: extract the agent program into internal/agentapp and add a dedicated
cmd/breeze-user-helper package so the user-helper embeds an asInvoker manifest
while breeze-agent keeps requireAdministrator. Each cmd wrapper keeps its own
var version so existing -X main.version ldflags are unchanged. VERSIONINFO is
preserved on both binaries (the #944 MSI "newer wins" behavior). Updates the
Makefile, ci.yml and release.yml build paths + winres.

Regression guards (both bidirectional): a PR-gating ci.yml job that cross-builds
both binaries and asserts user-helper=asInvoker / agent=requireAdministrator, and
the same assertion in release.yml on the shipped artifacts. The subsystem tests
now build the correct cmd packages by module path.

Live repro: device GBG-LT (OliveTech). Related helper-download sibling bug: #1878.
