+++ agent/cmd/breeze-user-helper/main.go
@@ -1,0 +1,22 @@
+// Command breeze-user-helper is the GUI-subsystem sibling of breeze-agent. It
+// builds the identical program (internal/agentapp) but lives in its own cmd/
+// package so it can embed an asInvoker manifest instead of the agent's
+// requireAdministrator one. This matters because the sessionbroker spawns the
+// user-helper into the logged-in user's session with that user's token; under
+// UAC a normal admin's token is filtered (non-elevated), so a
+// requireAdministrator manifest makes CreateProcessAsUser fail with
+// ERROR_ELEVATION_REQUIRED (740). The asInvoker manifest lets it launch in the
+// user's context. (The sessionbroker also spawns this same binary as SYSTEM for
+// desktop capture, where asInvoker is harmless; the user-token spawn path is the
+// one that requires it.) See cmd/breeze-user-helper/rsrc_windows_amd64.syso
+// (generated by `make build-winres`).
+package main
+
+import "github.com/breeze-rmm/agent/internal/agentapp"
+
+// version is overridden at build time via -ldflags "-X main.version=...".
+var version = "0.5.0"
+
+func main() {
+	agentapp.Main(version)
+}
