+++ agent/internal/agentapp/enroll_error.go
@@ -1,4 +1,4 @@
-package main
+package agentapp
 
 import (
 	"errors"
