+++ agent/internal/agentapp/enroll_error_test.go
@@ -1,4 +1,4 @@
-package main
+package agentapp
 
 import (
 	"bytes"
