+++ agent/internal/agentapp/reconcile_status_linux.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package main
+package agentapp
 
 import (
 	"fmt"
