+++ agent/internal/agentapp/reconcile_status_linux_test.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package main
+package agentapp
 
 import (
 	"os"
