+++ agent/internal/agentapp/service_helpers.go
@@ -1,4 +1,4 @@
-package main
+package agentapp
 
 import (
 	"context"
