+++ agent/internal/agentapp/service_seams_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package main
+package agentapp
 
 import "golang.org/x/sys/windows/svc"
