+++ agent/internal/agentapp/shutdown_timeout_test.go
@@ -1,4 +1,4 @@
-package main
+package agentapp
 
 import (
 	"testing"
