+++ agent/internal/agentapp/systemd_unit.go
@@ -1,4 +1,4 @@
-package main
+package agentapp
 
 import (
 	"fmt"
