+++ agent/internal/agentapp/systemd_unit_test.go
@@ -1,4 +1,4 @@
-package main
+package agentapp
 
 import (
 	"os"
