+++ agent/internal/agentapp/watchdog_bootstrap.go
@@ -1,4 +1,4 @@
-package main
+package agentapp
 
 import (
 	"bufio"
