+++ agent/internal/agentapp/watchdog_bootstrap_test.go
@@ -1,4 +1,4 @@
-package main
+package agentapp
 
 import (
 	"crypto/sha256"
